removed keyboard bug

+++ src/components/EndComponent.tsx
@@ -5,15 +5,18 @@ const EndComponent = ({
   answer,
   won,
   alreadyPlayed,
+  setCanPlay,
 }: {
   answer?: string;
   won?: boolean;
   alreadyPlayed: boolean;
+  setCanPlay: React.Dispatch<React.SetStateAction<boolean>>;
 }) => {
   useEffect(() => {
-    alreadyPlayed;
-    !alreadyPlayed &&
+    if (!alreadyPlayed) {
       localStorage.setItem("last-played", new Date().toISOString());
+      setCanPlay(false);
+    }
   }, []);
 
   if (alreadyPlayed) {
@@ -22,9 +25,9 @@ const EndComponent = ({
         initial={{ marginTop: 20, opacity: 0 }}
         animate={{ opacity: 1, marginTop: 0 }}
         transition={{ duration: 0.8, delay: 0.2 }}
-        className="fixed left-[50%] translate-x-[-50%] top-[50%] translate-y-[-50%] md:w-[30%] md:py-12 w-[80%]  flex justify-center flex-wrap py-3  border-border border-2 bg-background"
+        className="fixed left-[50%] top-[50%] flex w-[80%] translate-x-[-50%] translate-y-[-50%] flex-wrap  justify-center border-2 border-border bg-background  py-3 md:w-[30%] md:py-12"
       >
-        <h1 className="text-4xl mx-5 text-center">You already played today.</h1>
+        <h1 className="mx-5 text-center text-4xl">You already played today.</h1>
       </motion.div>
     );
   }
@@ -34,9 +37,9 @@ const EndComponent = ({
       initial={{ marginTop: 20, opacity: 0 }}
       animate={{ opacity: 1, marginTop: 0 }}
       transition={{ duration: 0.8, delay: 0.2 }}
-      className="fixed left-[50%] translate-x-[-50%] top-[50%] translate-y-[-50%] md:w-[30%] md:py-12 w-[80%] flex justify-center flex-wrap py-3  border-border border-2 bg-background"
+      className="fixed left-[50%] top-[50%] flex w-[80%] translate-x-[-50%] translate-y-[-50%] flex-wrap justify-center border-2 border-border bg-background  py-3 md:w-[30%] md:py-12"
     >
-      <h1 className="text-4xl font-bold text-center">
+      <h1 className="text-center text-4xl font-bold">
         You{" "}
         {won ? (
           <span className="text-green-500">won.</span>
@@ -45,12 +48,12 @@ const EndComponent = ({
         )}
       </h1>
       {!won ? (
-        <div className="mt-3 text-xl text-center">
+        <div className="mt-3 text-center text-xl">
           Correct answer was <span className="font-semibold">{answer}</span>.{" "}
           Come back tomorrow to try your luck again.
         </div>
       ) : (
-        <div className="mt-3 text-center  px-5 text-xl">
+        <div className="mt-3 px-5  text-center text-xl">
           Thanks for playing. Come back tomorrow to play once more.{" "}
         </div>
       )}
